Clean up architecture: single @Mod, remove forge/, move core to submodule
- Rename submodule branch interface -> core, update .gitmodules
- Remove forge/ package entirely, merge into ae2interface/
- Single @Mod (ae2webintegration) combining former -core and -interface
- Fix shadow() -> shadowImplementation() + usesShadowedDependencies
- Move web resources (html, favicon) to web-engine submodule
- Update mcmod.info to single mod entry
- Fix mixin references: GameProfile -> PlayerIdentity/UUID, AEControllerProfile -> AEControllerUUID
