+++ src/main/java/pl/kuba6000/ae2webintegration/ae2interface/AE2WebIntegration.java
@@ -3,12 +3,17 @@ package pl.kuba6000.ae2webintegration.ae2interface;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import com.mojang.authlib.GameProfile;
+
 import appeng.me.cache.SecurityCache;
 import cpw.mods.fml.common.Mod;
+import cpw.mods.fml.common.SidedProxy;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import cpw.mods.fml.common.event.FMLServerStartedEvent;
 import cpw.mods.fml.common.event.FMLServerStartingEvent;
+import cpw.mods.fml.common.event.FMLServerStoppingEvent;
 import pl.kuba6000.ae2webintegration.Tags;
 import pl.kuba6000.ae2webintegration.ae2interface.implementations.AE;
 import pl.kuba6000.ae2webintegration.core.api.IAEWebInterface;
@@ -16,33 +21,53 @@ import pl.kuba6000.ae2webintegration.core.api.IAEWebInterface;
 @Mod(
     modid = AE2WebIntegration.MODID,
     version = Tags.VERSION,
-    name = "AE2WebIntegration-Interface",
+    name = "AE2 Web Integration",
     acceptedMinecraftVersions = "[1.7.10]",
     acceptableRemoteVersions = "*")
 public class AE2WebIntegration {
 
-    public static final String MODID = "ae2webintegration-interface";
+    public static final String MODID = "ae2webintegration";
     public static final Logger LOG = LogManager.getLogger(MODID);
 
+    @SidedProxy(
+        clientSide = "pl.kuba6000.ae2webintegration.ae2interface.ClientProxy",
+        serverSide = "pl.kuba6000.ae2webintegration.ae2interface.CommonProxy")
+    public static CommonProxy proxy;
+
     @Mod.EventHandler
-    public void preInit(FMLPreInitializationEvent event) {}
+    public void preInit(FMLPreInitializationEvent event) {
+        proxy.preInit(event);
+    }
 
     @Mod.EventHandler
     public void init(FMLInitializationEvent event) {
+        proxy.init(event);
         IAEWebInterface.getInstance()
             .initAEInterface(AE.instance);
     }
 
     @Mod.EventHandler
     public void postInit(FMLPostInitializationEvent event) {
+        proxy.postInit(event);
         SecurityCache.registerOpPlayer(
-            IAEWebInterface.getInstance()
-                .getAEWebGameProfile());
+            new GameProfile(
+                IAEWebInterface.getInstance()
+                    .getAEWebUUID(),
+                "AE2CONTROLLER"));
     }
 
     @Mod.EventHandler
     public void serverStarting(FMLServerStartingEvent event) {
-
+        proxy.serverStarting(event);
     }
 
+    @Mod.EventHandler
+    public void serverStarted(FMLServerStartedEvent event) {
+        proxy.serverStarted(event);
+    }
+
+    @Mod.EventHandler
+    public void serverStopping(FMLServerStoppingEvent event) {
+        proxy.serverStopping(event);
+    }
 }
