+++ src/main/resources/mcmod.info
@@ -1,26 +1,8 @@
 {
 	"modListVersion": 2,
 	"modList": [{
-		"modid": "ae2webintegration-core",
-		"name": "AE2WebIntegration-Core",
-		"description": "AE2 Web Integration mod",
-		"version": "${modVersion}",
-		"mcversion": "*",
-		"url": "https://github.com/kuba6000/AE2-Web-Integration",
-		"updateUrl": "",
-		"authorList": ["kuba6000"],
-		"credits": "",
-		"logoFile": "",
-		"screenshots": [],
-		"parent": "",
-		"requiredMods": [],
-		"dependencies": [],
-		"dependants": [],
-		"useDependencyInformation": false
-	},
-	{
-    "modid": "ae2webintegration-interface",
-		"name": "AE2WebIntegration-Interface",
+		"modid": "ae2webintegration",
+		"name": "AE2 Web Integration",
 		"description": "AE2 Web Integration mod",
 		"version": "${modVersion}",
 		"mcversion": "${minecraftVersion}",
